+++ dhis-2/dhis-test-integration/src/test/java/org/hisp/dhis/dxf2/deprecated/tracker/aggregates/TrackedEntityAttributesAggregateAclTest.java
@@ -35,8 +35,8 @@ import java.util.List;
 import java.util.Map;
 
 import org.hisp.dhis.common.CodeGenerator;
-import org.hisp.dhis.dxf2.TrackerTest;
 import org.hisp.dhis.dxf2.deprecated.tracker.TrackedEntityInstanceParams;
+import org.hisp.dhis.dxf2.deprecated.tracker.TrackerTest;
 import org.hisp.dhis.dxf2.deprecated.tracker.trackedentity.TrackedEntityInstance;
 import org.hisp.dhis.dxf2.deprecated.tracker.trackedentity.TrackedEntityInstanceService;
 import org.hisp.dhis.trackedentity.TrackedEntityQueryParams;
